Allow writing to to readonly fields
Fixes #13237

On both .NET Core and .NET Framework, the expression builder APIs do not allow assignment to IsInitOnly fields. However, on .NET Core, if this check is bypassed, then the compiler will happily compile the expression tree, which results in performant code for setting readonly fields. We should decide if we're comfortable doing this.

On .NET Framework, the compiler will not compile the expression tree, so if not running on Core, we drop down to calling FieldInfo.SetValue, which both boxes and is on the order of 1000 times slower, but has always worked and so should be safe to do.